refactor [training] - group update methods on module
